Move IONKEYS to the root
